refactor: isolate group norm fallback
